redefine typedef

+++ src/util/json/cJSON.h
@@ -23,11 +23,13 @@
 #ifndef cJSON__h
 #define cJSON__h
 
+#include <stdint.h>
 
-typedef int int32;
-typedef unsigned int uint32;
-typedef long long int64;
-typedef unsigned long long uint64;
+typedef int32_t int32;
+typedef uint32_t uint32;
+typedef int64_t int64;
+typedef uint64_t uint64;
+
 
 #ifdef __cplusplus
 extern "C"
